Create mutation_nomenclature 0.1
This is the first iteration of mutation_nomenclature, It can only take inputs from user and show what kind of reference sequence it is. In the future, I wish for it to be able to give a full description of the mutation. To expand upon this project, I would like for it to output a correct nomenclature when given a description and/or when given a before and after pdb file.